# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearUsuarioDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActualizarUsuarioDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/addresses":{"post":{"operationId":"AddressesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearDireccionDto"}}}},"responses":{"201":{"description":""}},"tags":["Addresses"]},"get":{"operationId":"AddressesController_findAll","parameters":[{"name":"userId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Addresses"]}},"/api/addresses/{id}":{"get":{"operationId":"AddressesController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Addresses"]},"patch":{"operationId":"AddressesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActualizarDireccionDto"}}}},"responses":{"200":{"description":""}},"tags":["Addresses"]},"delete":{"operationId":"AddressesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Addresses"]}},"/api/cities":{"post":{"operationId":"CitiesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearCiudadDto"}}}},"responses":{"201":{"description":""}},"tags":["Cities"]},"get":{"operationId":"CitiesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Cities"]}},"/api/cities/{id}":{"get":{"operationId":"CitiesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cities"]},"patch":{"operationId":"CitiesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActualizarCiudadDto"}}}},"responses":{"200":{"description":""}},"tags":["Cities"]},"delete":{"operationId":"CitiesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cities"]}},"/api/carriers/quote":{"post":{"description":"Obtiene el costo de envío usando códigos de Chilexpress directamente  Para obtener los códigos correctos, usar primero GET /geo/chilexpress/regions y luego GET /geo/chilexpress/coverage-areas","operationId":"CarriersController_quote","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolicitudCotizacionDto"},"examples":{"Usando countyCode":{"value":{"originCountyCode":"STGO","destinationCountyCode":"PROV","package":{"weight":"2.5","height":"15","width":"25","length":"35"},"productType":3,"contentType":1,"declaredWorth":"25000","deliveryTime":0}}}}}},"responses":{"200":{"description":"Cotización exitosa con opciones de servicio","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RespuestaCotizacionDto"}}}}},"summary":"Cotizar envío","tags":["Carriers"]}},"/api/carriers":{"post":{"operationId":"CarriersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearTransportistaDto"}}}},"responses":{"201":{"description":""}},"summary":"Crear un nuevo carrier","tags":["Carriers"]},"get":{"operationId":"CarriersController_findAll","parameters":[],"responses":{"200":{"description":""}},"summary":"Listar todos los carriers","tags":["Carriers"]}},"/api/carriers/{id}":{"get":{"operationId":"CarriersController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Obtener carrier por ID","tags":["Carriers"]},"patch":{"operationId":"CarriersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActualizarTransportistaDto"}}}},"responses":{"200":{"description":""}},"summary":"Actualizar carrier","tags":["Carriers"]},"delete":{"operationId":"CarriersController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Eliminar carrier","tags":["Carriers"]}},"/api/geo/cl/regions":{"get":{"description":"Obtiene regiones, provincias y comunas","operationId":"GeoController_getRegions","parameters":[],"responses":{"200":{"description":""}},"summary":"Obtener regiones de Chile","tags":["Geo"]}},"/api/geo/chilexpress/regions":{"get":{"description":"Obtiene las regiones disponibles en Chilexpress","operationId":"GeoController_getChilexpressRegions","parameters":[],"responses":{"200":{"description":""}},"summary":"Obtener regiones de Chilexpress","tags":["Geo"]}},"/api/geo/chilexpress/coverage-areas":{"get":{"description":"Obtiene las comunas/áreas de cobertura para una región específica","operationId":"GeoController_getChilexpressCoverageAreas","parameters":[{"name":"regionCode","required":true,"in":"query","description":"Código de región de Chilexpress","schema":{"type":"string"}},{"name":"type","required":false,"in":"query","description":"Tipo de consulta (0 = todas las áreas)","schema":{"default":0,"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Obtener áreas de cobertura de Chilexpress","tags":["Geo"]}},"/api/geo/chilexpress/county-by-name/{communeName}":{"get":{"description":"Busca un código de Chilexpress por nombre de comuna","operationId":"GeoController_findChilexpressCountyByName","parameters":[{"name":"communeName","required":true,"in":"path","description":"Nombre de la comuna a buscar","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Buscar county code por nombre de comuna","tags":["Geo"]}},"/api/deliveries/create-from-payment":{"post":{"description":"Endpoint llamado automáticamente por el microservicio de Pagos cuando un pago se completa. Crea el envío con estado \"Preparando\" y genera el tracking number.","operationId":"DeliveriesController_createFromPayment","parameters":[],"requestBody":{"required":true,"description":"Datos del pago y envío","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearEnvioDesdePagoDto"},"examples":{"example1":{"summary":"Ejemplo completo","value":{"paymentId":"pay_abc123","cartId":"cart_xyz789","userId":"user_456","sellerId":"seller_789","totalAmount":950000,"items":[{"productId":"prod_12345","name":"iPhone 14 Pro 256GB","quantity":1,"price":899990}],"package":{"weight":0.5,"length":20,"width":15,"height":10},"shippingInfo":{"originAddressId":"addr_origin_123","destinationAddressId":"addr_dest_456","carrierName":"Chilexpress","serviceType":"PRIORITARIO","estimatedCost":8812},"declaredWorth":50000,"notes":"Entregar en horario de oficina"}}}}}},"responses":{"201":{"description":"Envío creado exitosamente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RespuestaEnvioDto"},"example":{"trackingNumber":"ENV-1734480000000-A3B7F9","status":"Preparando","paymentId":"pay_abc123","cartId":"cart_xyz789","userId":"user_456","sellerId":"seller_789","carrierName":"Chilexpress","serviceType":"PRIORITARIO","estimatedCost":8812,"currency":"CLP","originAddressId":"addr_origin_123","destinationAddressId":"addr_dest_456","items":[{"productId":"prod_12345","name":"iPhone 14 Pro 256GB","quantity":1,"price":899990}],"estimatedDeliveryDate":"2024-12-18T10:00:00Z","createdAt":"2024-12-17T15:30:00Z","message":"Envío creado exitosamente. El vendedor ha sido notificado."}}}},"400":{"description":"Datos inválidos o envío ya existe para este pago"},"500":{"description":"Error al crear el envío"}},"summary":"Crear envío desde pago completado","tags":["deliveries"]}},"/api/deliveries/create":{"post":{"description":"Crea un envío desde el frontend sin necesidad de un pago previo. Usado en el flujo de cotización.","operationId":"DeliveriesController_createDirectly","parameters":[],"requestBody":{"required":true,"description":"Datos del envío","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearEnvioDirectoDto"},"examples":{"example1":{"summary":"Ejemplo de creación directa","value":{"userId":"user_456","sellerId":"seller_789","cartId":"cart_xyz789","items":[{"productId":"prod_12345","name":"iPhone 14 Pro 256GB","quantity":1,"price":899990}],"package":{"weight":0.5,"length":20,"width":15,"height":10},"shippingInfo":{"originAddressId":"addr_origin_123","destinationAddressId":"addr_dest_456","carrierName":"Chilexpress","serviceType":"PRIORITARIO","estimatedCost":8812},"declaredWorth":50000,"notes":"Entregar en horario de oficina"}}}}}},"responses":{"201":{"description":"Envío creado exitosamente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RespuestaEnvioDto"}}}},"400":{"description":"Datos inválidos"}},"summary":"Crear envío directamente","tags":["deliveries"]}},"/api/deliveries/user/{userId}":{"get":{"description":"Obtiene todos los envíos asociados a un usuario específico, ordenados por fecha de creación descendente.","operationId":"DeliveriesController_findByUserId","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Lista de envíos del usuario","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RespuestaEnvioDto"}}}}},"404":{"description":"Usuario sin envíos"}},"summary":"Listar envíos del usuario","tags":["deliveries"]}},"/api/deliveries/tracking/{trackingNumber}":{"get":{"description":"Obtiene los detalles completos de un envío usando su número de seguimiento (tracking number).","operationId":"DeliveriesController_findByTrackingNumber","parameters":[{"name":"trackingNumber","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Envío encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RespuestaEnvioDto"}}}},"404":{"description":"Envío no encontrado con ese número de tracking"}},"summary":"Buscar envío por número de tracking","tags":["deliveries"]}},"/api/deliveries":{"get":{"operationId":"DeliveriesController_findAll","parameters":[],"responses":{"200":{"description":"Lista de envíos"}},"summary":"Listar todos los envíos","tags":["deliveries"]}},"/api/deliveries/{id}":{"get":{"operationId":"DeliveriesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Envío encontrado"},"404":{"description":"Envío no encontrado"}},"summary":"Obtener envío por ID","tags":["deliveries"]},"patch":{"operationId":"DeliveriesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryDto"}}}},"responses":{"200":{"description":"Envío actualizado"},"404":{"description":"Envío no encontrado"}},"summary":"Actualizar envío","tags":["deliveries"]},"delete":{"operationId":"DeliveriesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Envío eliminado"}},"summary":"Eliminar envío","tags":["deliveries"]}}},"info":{"title":"PulgaShop - Microservicio Envíos","description":"API para direcciones y transportistas","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CrearUsuarioDto":{"type":"object","properties":{"nombre":{"type":"string"},"apellido":{"type":"string"},"correo":{"type":"string"},"contrasena":{"type":"string"}},"required":["nombre","apellido","correo","contrasena"]},"ActualizarUsuarioDto":{"type":"object","properties":{}},"CrearDireccionDto":{"type":"object","properties":{"calle":{"type":"string"},"numero":{"type":"string"},"comuna_id":{"type":"string"},"codigo_comuna":{"type":"string","description":"Código Chilexpress"},"provincia":{"type":"string"},"region_id":{"type":"string"},"codigo_postal":{"type":"string"},"referencias":{"type":"string"},"usuario_id":{"type":"string"}},"required":["calle","numero","comuna_id","region_id","usuario_id"]},"ActualizarDireccionDto":{"type":"object","properties":{"calle":{"type":"string"},"numero":{"type":"string"},"comuna_id":{"type":"string"},"codigo_comuna":{"type":"string","description":"Código Chilexpress"},"provincia":{"type":"string"},"region_id":{"type":"string"},"codigo_postal":{"type":"string"},"referencias":{"type":"string"},"usuario_id":{"type":"string"}}},"CrearCiudadDto":{"type":"object","properties":{"nombre":{"type":"string","example":"Santo Domingo"},"codigo_postal":{"type":"string","example":"2720000"}},"required":["nombre","codigo_postal"]},"ActualizarCiudadDto":{"type":"object","properties":{}},"InfoPaqueteDto":{"type":"object","properties":{"peso":{"type":"string","example":"2.5","description":"Peso del paquete en kilogramos (separado por punto)"},"alto":{"type":"string","example":"15","description":"Alto del paquete en centímetros"},"ancho":{"type":"string","example":"25","description":"Ancho del paquete en centímetros"},"largo":{"type":"string","example":"35","description":"Largo del paquete en centímetros"}},"required":["peso","alto","ancho","largo"]},"SolicitudCotizacionDto":{"type":"object","properties":{"comuna_origen_id":{"type":"string","example":"13101","description":"Código DPA de la comuna de origen (ej: 13101 = Santiago, 05109 = Viña del Mar). Usar solo si no se envía codigo_cobertura_origen"},"comuna_destino_id":{"type":"string","example":"05109","description":"Código DPA de la comuna de destino (ej: 13101 = Santiago, 05109 = Viña del Mar). Usar solo si no se envía codigo_cobertura_destino"},"codigo_cobertura_origen":{"type":"string","example":"STGO","description":"Código de cobertura de Chilexpress para origen (obtenido desde /geo/chilexpress/coverage-areas). Tiene prioridad sobre comuna_origen_id"},"codigo_cobertura_destino":{"type":"string","example":"VAP","description":"Código de cobertura de Chilexpress para destino (obtenido desde /geo/chilexpress/coverage-areas). Tiene prioridad sobre comuna_destino_id"},"paquete":{"description":"Información del paquete a enviar","allOf":[{"$ref":"#/components/schemas/InfoPaqueteDto"}]},"tipo_producto":{"type":"number","example":3,"description":"Tipo de producto: 1 = Documento, 3 = Encomienda","enum":[1,3]},"tipo_contenido":{"type":"number","example":1,"description":"Tipo de contenido del paquete"},"valor_declarado":{"type":"string","example":"25000","description":"Valor declarado del paquete en pesos chilenos"},"tiempo_entrega":{"type":"number","example":0,"description":"Servicios de entrega: 0 = Todos, 1 = Prioritarios, 2 = No prioritarios, 3 = De devolución","enum":[0,1,2,3],"default":0}},"required":["paquete","tipo_producto","tipo_contenido","valor_declarado","tiempo_entrega"]},"OpcionServicioDto":{"type":"object","properties":{"nombre_servicio":{"type":"string","example":"PRIORITARIO","description":"Nombre del servicio de envío"},"codigo_servicio":{"type":"string","example":"PRI","description":"Código del servicio"},"valor_servicio":{"type":"number","example":8500,"description":"Costo del servicio en pesos chilenos"},"tiempo_entrega":{"type":"string","example":"1 día hábil","description":"Tiempo estimado de entrega"},"disponible":{"type":"boolean","example":true,"description":"Indica si el servicio está disponible"}},"required":["nombre_servicio","codigo_servicio","valor_servicio","tiempo_entrega","disponible"]},"RespuestaCotizacionDto":{"type":"object","properties":{"codigo_estado":{"type":"number","example":0,"description":"Código de estado de la respuesta (0 = éxito)"},"descripcion_estado":{"type":"string","example":"Cotización exitosa","description":"Descripción del estado"},"opciones_servicio":{"description":"Lista de opciones de servicio disponibles","type":"array","items":{"$ref":"#/components/schemas/OpcionServicioDto"}},"comuna_origen_id":{"type":"string","example":"13101","description":"Código de comuna de origen utilizado"},"comuna_destino_id":{"type":"string","example":"05109","description":"Código de comuna de destino utilizado"},"codigo_cobertura_origen":{"type":"string","example":"SCL","description":"Código de cobertura Chilexpress de origen"},"codigo_cobertura_destino":{"type":"string","example":"VAP","description":"Código de cobertura Chilexpress de destino"},"error":{"type":"string","description":"Mensaje de error si la cotización falló"},"datos_crudos":{"type":"object","description":"Datos adicionales de la respuesta de Chilexpress"}},"required":["codigo_estado","descripcion_estado","opciones_servicio","comuna_origen_id","comuna_destino_id","codigo_cobertura_origen","codigo_cobertura_destino","error","datos_crudos"]},"CrearTransportistaDto":{"type":"object","properties":{"nombre":{"type":"string","example":"DHL"},"credenciales":{"type":"object","example":{"X-API-AppKey":"Luke Skywalker","X-API-AppToken":"R2D2"}},"disponible":{"type":"boolean","example":true}},"required":["nombre","credenciales","disponible"]},"ActualizarTransportistaDto":{"type":"object","properties":{}},"ArticuloCarritoDto":{"type":"object","properties":{"producto_id":{"type":"string","example":"prod_12345","description":"ID del producto"},"nombre":{"type":"string","example":"iPhone 14 Pro","description":"Nombre del producto"},"cantidad":{"type":"number","example":1,"description":"Cantidad"},"precio":{"type":"number","example":899990,"description":"Precio unitario"}},"required":["producto_id","nombre","cantidad","precio"]},"InfoEnvioDto":{"type":"object","properties":{"origen_direccion_id":{"type":"string","example":"addr_12345","description":"ID de dirección de origen"},"destino_direccion_id":{"type":"string","example":"addr_67890","description":"ID de dirección de destino"},"nombre_transportista":{"type":"string","example":"Chilexpress","description":"Nombre del carrier"},"tipo_servicio":{"type":"string","example":"PRIORITARIO","description":"Tipo de servicio"},"costo_estimado":{"type":"number","example":8812,"description":"Costo estimado del envío"}},"required":["origen_direccion_id","destino_direccion_id","nombre_transportista","tipo_servicio","costo_estimado"]},"CrearEnvioDesdePagoDto":{"type":"object","properties":{"pago_id":{"type":"string","example":"pay_abc123","description":"ID del pago completado"},"carrito_id":{"type":"string","example":"cart_xyz789","description":"ID del carrito"},"usuario_id":{"type":"string","example":"user_456","description":"ID del usuario comprador"},"vendedor_id":{"type":"string","example":"seller_789","description":"ID del vendedor"},"monto_total":{"type":"number","example":950000,"description":"Monto total del pago"},"articulo_carrito":{"description":"Items del carrito","type":"array","items":{"$ref":"#/components/schemas/ArticuloCarritoDto"}},"paquete":{"description":"Información del paquete","allOf":[{"$ref":"#/components/schemas/InfoPaqueteDto"}]},"informacion_envio":{"description":"Información de envío (previamente cotizada)","allOf":[{"$ref":"#/components/schemas/InfoEnvioDto"}]},"valor_declarado":{"type":"number","example":50000,"description":"Valor declarado para seguro"},"notas":{"type":"string","example":"Envío de compra online - Manejar con cuidado","description":"Notas adicionales"}},"required":["pago_id","carrito_id","usuario_id","vendedor_id","monto_total","articulo_carrito","paquete","informacion_envio"]},"RespuestaEnvioDto":{"type":"object","properties":{"numero_seguimiento":{"type":"string","example":"ENV-1734480000000-A3B7F9","description":"Número de tracking único"},"estado":{"type":"string","example":"Preparando","enum":["Preparando","EnTransito","Entregado","Cancelado","Devuelto"],"description":"Estado actual del envío"},"pago_id":{"type":"string","example":"pay_abc123","description":"ID del pago"},"carrito_id":{"type":"string","example":"cart_xyz789","description":"ID del carrito"},"usuario_id":{"type":"string","example":"user_456","description":"ID del usuario"},"vendedor_id":{"type":"string","example":"seller_789","description":"ID del vendedor"},"nombre_transportista":{"type":"string","example":"Chilexpress","description":"Nombre del carrier"},"tipo_servicio":{"type":"string","example":"PRIORITARIO","description":"Tipo de servicio"},"costo_estimado":{"type":"number","example":8812,"description":"Costo estimado"},"moneda":{"type":"string","example":"CLP","description":"Moneda"},"origen_direccion_id":{"type":"string","example":"addr_12345","description":"ID dirección origen"},"destino_direccion_id":{"type":"string","example":"addr_67890","description":"ID dirección destino"},"articulos":{"example":[{"producto_id":"prod_1","nombre":"iPhone 14","cantidad":1,"precio":899990}],"description":"Items incluidos en el envío","type":"array","items":{"type":"string"}},"fecha_entrega_estimada":{"format":"date-time","type":"string","example":"2024-12-20T10:00:00Z","description":"Fecha estimada de entrega"},"creado_en":{"format":"date-time","type":"string","example":"2024-12-17T15:30:00Z","description":"Fecha de creación"},"mensaje":{"type":"string","example":"Envío creado exitosamente. El vendedor ha sido notificado.","description":"Mensaje informativo"}},"required":["numero_seguimiento","estado","pago_id","carrito_id","usuario_id","vendedor_id","nombre_transportista","tipo_servicio","costo_estimado","moneda","origen_direccion_id","destino_direccion_id","articulos","fecha_entrega_estimada","creado_en","mensaje"]},"CrearEnvioDirectoDto":{"type":"object","properties":{"usuario_id":{"type":"string","example":"user_456","description":"ID del usuario que crea el envío"},"vendedor_id":{"type":"string","example":"seller_789","description":"ID del vendedor"},"carrito_id":{"type":"string","example":"cart_xyz789","description":"ID del carrito (referencia)"},"articulo_carrito":{"description":"Items del envío","type":"array","items":{"$ref":"#/components/schemas/ArticuloCarritoDto"}},"paquete":{"description":"Información del paquete","allOf":[{"$ref":"#/components/schemas/InfoPaqueteDto"}]},"informacion_envio":{"description":"Información de envío (previamente cotizada)","allOf":[{"$ref":"#/components/schemas/InfoEnvioDto"}]},"valor_declarado":{"type":"number","example":50000,"description":"Valor declarado para seguro"},"notas":{"type":"string","example":"Entregar en horario de oficina","description":"Notas adicionales"}},"required":["usuario_id","vendedor_id","carrito_id","articulo_carrito","paquete","informacion_envio"]},"PackageDto":{"type":"object","properties":{"weight":{"type":"string","example":"16","description":"Package weight (kg)"},"height":{"type":"string","example":"10","description":"Package height (cm)"},"width":{"type":"string","example":"20","description":"Package width (cm)"},"length":{"type":"string","example":"30","description":"Package length (cm)"}},"required":["weight","height","width","length"]},"DeliveryDto":{"type":"object","properties":{"originCountyCode":{"type":"string","example":"STGO","description":"Origin commune DPA code (API Gobierno)"},"destinationCountyCode":{"type":"string","example":"PROV","description":"Destination commune DPA code (API Gobierno)"},"package":{"$ref":"#/components/schemas/PackageDto"},"productType":{"type":"number","example":3,"description":"Product type code"},"contentType":{"type":"number","example":1,"description":"Content type code"},"declaredWorth":{"type":"string","example":"2333","description":"Declared worth (currency amount)"},"deliveryTime":{"type":"number","example":0,"description":"Delivery time preference (0 = standard)"}},"required":["originCountyCode","destinationCountyCode","package","productType","contentType","declaredWorth","deliveryTime"]}}}}